requirements.txt
alembic
cryptography==43.0.1
Flask==3.0.3
flask_mail==0.10.0
flask_session
flask_wtf==1.2.1
gunicorn
itsdangerous==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
Werkzeug==3.0.4
